requirements_windows.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
GPUtil==1.4.0
idna==3.4
importlib-metadata==5.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.2
numba==0.56.4
numpy==1.23.5
opencv-python==4.7.0.68
packaging==22.0
Pillow==9.3.0
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
